Counter doesn't go negative now
Counter used to start counting to negative numbers if the time
to elapse was set to 0.
I added extra code inside the if and else if statements to make
sure to modify the count values only if it is correct.
Implemented if statement for pomoMinus that prevents the minutes
going negative on the toggleMenu.